@@ -1,9 +1,11 @@
-package com.github.phonenumbermanager.validator;
+package com.github.phonenumbermanager.annotation;
 
 import java.lang.annotation.*;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
+
+import com.github.phonenumbermanager.validator.PhoneNumberConstraintValidator;
 
 /**
  * 验证联系方式注解
@@ -17,7 +19,7 @@ import javax.validation.Payload;
 @Documented
 @Constraint(validatedBy = {PhoneNumberConstraintValidator.class})
 public @interface PhoneNumberValidator {
-    String message() default "{com.github.phonenumbermanager.validator.PhoneNumberValidator.message}";
+    String message() default "{com.github.phonenumbermanager.annotation.PhoneNumberValidator.message}";
 
     Class<?>[] groups() default {};
 
