@@ -26,7 +26,11 @@ public enum EmploymentStatusEnum {
     private final int value;
     private final String description;
 
-    public static EmploymentStatusEnum getEnum(String description) {
+    public static EmploymentStatusEnum valueOf(int value) {
+        return Arrays.stream(values()).filter(e -> value == e.getValue()).findFirst().orElse(null);
+    }
+
+    public static EmploymentStatusEnum descriptionOf(String description) {
         return Arrays.stream(values()).filter(e -> e.getDescription().equals(description)).findFirst().orElse(null);
     }
 }
