@@ -7,30 +7,33 @@ import com.fasterxml.jackson.annotation.JsonValue;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.ToString;
 
 /**
- * 数据类别枚举
+ * 异步导入、导出状态枚举
  *
  * @author 廿二月的天
  */
-@AllArgsConstructor
 @Getter
-public enum FieldTypeEnum {
+@AllArgsConstructor
+@ToString
+public enum ImportOrExportStatusEnum {
     /**
      *
      */
-    UNKNOWN(0, "未知类型"), BOOLEAN(1, "布尔类型"), STRING(2, "字符串类型"), NUMBER(3, "数值类型"), SYSTEM_USER(4, "系统用户类型");
+    START(0, "开始运行"), HANDLING(1, "处理中"), HANDLED(2, "处理完成"), DOWNLOAD(3, "下载"), UPLOAD(4, "上传"), DONE(5, "运行成功"),
+    FAILED(6, "运行失败");
 
     @EnumValue
     @JsonValue
     private final int value;
     private final String description;
 
-    public static FieldTypeEnum valueOf(int value) {
+    public static ImportOrExportStatusEnum valueOf(int value) {
         return Arrays.stream(values()).filter(e -> value == e.getValue()).findFirst().orElse(null);
     }
 
-    public static FieldTypeEnum descriptionOf(String description) {
+    public static ImportOrExportStatusEnum descriptionOf(String description) {
         return Arrays.stream(values()).filter(e -> e.getDescription().equals(description)).findFirst().orElse(null);
     }
 }
