@@ -27,7 +27,11 @@ public enum PoliticalStatusEnum {
     private final int value;
     private final String description;
 
-    public static PoliticalStatusEnum getEnum(String description) {
+    public static PoliticalStatusEnum valueOf(int value) {
+        return Arrays.stream(values()).filter(e -> value == e.getValue()).findFirst().orElse(null);
+    }
+
+    public static PoliticalStatusEnum descriptionOf(String description) {
         return Arrays.stream(values()).filter(e -> e.getDescription().equals(description)).findFirst().orElse(null);
     }
 }
