@@ -1,38 +1,31 @@
 package com.github.phonenumbermanager.vo;
 
 import java.io.Serializable;
-import java.util.List;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
 
 import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
 /**
- * 列表框视图对象
+ * 计算视图对象
  *
  * @author 廿二月的天
  */
-@ApiModel("列表框视图对象")
 @Data
 @NoArgsConstructor
-@Accessors(chain = true)
 @EqualsAndHashCode
-public class SelectListVo implements Serializable {
+@Accessors(chain = true)
+@ApiModel("计算视图对象")
+public class ComputedVO implements Serializable {
+    @ApiModelProperty("需要获取的单位编号集合")
     @JsonSerialize(using = ToStringSerializer.class)
-    private Long id;
-    private String name;
-    private String title;
-    private String label;
-    @JsonSerialize(using = ToStringSerializer.class)
-    private Long value;
-    private Boolean isLeaf;
-    private Boolean loading;
-    private Boolean isSubordinate;
-    private Boolean isGrandsonLevel;
-    private List<SelectListVo> children;
+    private Long[] companyIds;
+    @ApiModelProperty("柱状图表类型参数")
+    private Boolean barChartTypeParam;
 }
