@@ -1,31 +1,38 @@
 package com.github.phonenumbermanager.vo;
 
 import java.io.Serializable;
+import java.util.List;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
 
 import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
 /**
- * 计算视图对象
+ * 列表框视图对象
  *
  * @author 廿二月的天
  */
+@ApiModel("列表框视图对象")
 @Data
 @NoArgsConstructor
-@EqualsAndHashCode
 @Accessors(chain = true)
-@ApiModel("计算视图对象")
-public class ComputedVo implements Serializable {
-    @ApiModelProperty("需要获取的单位编号集合")
+@EqualsAndHashCode
+public class SelectListVO implements Serializable {
     @JsonSerialize(using = ToStringSerializer.class)
-    private Long[] companyIds;
-    @ApiModelProperty("柱状图表类型参数")
-    private Boolean barChartTypeParam;
+    private Long id;
+    private String name;
+    private String title;
+    private String label;
+    @JsonSerialize(using = ToStringSerializer.class)
+    private Long value;
+    private Boolean isLeaf;
+    private Boolean loading;
+    private Boolean isSubordinate;
+    private Boolean isGrandsonLevel;
+    private List<SelectListVO> children;
 }
