@@ -2,11 +2,16 @@ import {
   communityResident,
   communityResidentBatch,
   communityResidentDownloadExcel,
+  communityResidentExportExcel,
   communityResidentImportExcel,
 } from '@/services/api';
 import request from '@config/request';
 
-/** 获取社区居民列表 */
+/**
+ * 获取社区居民列表
+ * @param params
+ * @param options
+ */
 export async function queryCommunityResidentList(params?: any, options?: Record<string, any>) {
   return request.get<API.DataList<API.CommunityResident> & API.ResponseException>(
     communityResident,
@@ -17,7 +22,11 @@ export async function queryCommunityResidentList(params?: any, options?: Record<
   );
 }
 
-/** 获取社区居民列表 */
+/**
+ * 获取社区居民列表
+ * @param id
+ * @param options
+ */
 export async function queryCommunityResident(id: number, options?: Record<string, any>) {
   return request.get<API.Data<API.CommunityResident> & API.ResponseException>(
     `${communityResident}/${id}`,
@@ -27,7 +36,12 @@ export async function queryCommunityResident(id: number, options?: Record<string
   );
 }
 
-/** 单独字段修改社区居民 */
+/**
+ * 单独字段修改社区居民
+ * @param id
+ * @param data
+ * @param options
+ */
 export async function patchCommunityResident(
   id: number,
   data: API.CommunityResident,
@@ -42,7 +56,11 @@ export async function patchCommunityResident(
   );
 }
 
-/** 添加社区居民处理 */
+/**
+ * 添加社区居民处理
+ * @param data
+ * @param options
+ */
 export async function createCommunityResident(
   data: API.CommunityResident,
   options?: Record<string, any>,
@@ -53,7 +71,12 @@ export async function createCommunityResident(
   });
 }
 
-/** 修改社区居民处理 */
+/**
+ * 修改社区居民处理
+ * @param id
+ * @param data
+ * @param options
+ */
 export async function modifyCommunityResident(
   id: number,
   data: API.CommunityResident,
@@ -65,14 +88,22 @@ export async function modifyCommunityResident(
   });
 }
 
-/** 删除社区居民 */
+/**
+ * 删除社区居民
+ * @param id
+ * @param options
+ */
 export async function removeCommunityResident(id: number, options?: Record<string, any>) {
   return request.delete<API.ResponseSuccess & API.ResponseException>(`${communityResident}/${id}`, {
     ...(options || {}),
   });
 }
 
-/** 批量操作社区居民 */
+/**
+ * 批量操作社区居民
+ * @param data
+ * @param options
+ */
 export async function batchCommunityResident<T>(
   data: API.BatchRUD<T[]>,
   options?: Record<string, any>,
@@ -83,21 +114,60 @@ export async function batchCommunityResident<T>(
   });
 }
 
-/** 上传表格 */
+/**
+ * 上传表格
+ * @param data
+ * @param importId
+ * @param optionsObject
+ */
 export async function uploadCommunityResidentExcel(
-  formData: FormData,
+  data?: FormData,
+  importId?: number,
+  optionsObject?: Record<string, any>,
+) {
+  let options = optionsObject || {};
+  if (typeof data !== 'undefined') {
+    options = { ...optionsObject, requestType: 'form', data };
+  }
+  if (typeof importId !== 'undefined') {
+    options = { ...options, params: { importId } };
+  }
+  return request.post<API.ImportFileProgress & API.ResponseException>(
+    communityResidentImportExcel,
+    options,
+  );
+}
+
+/**
+ * 生成 Excel 文件
+ * @param exportId
+ * @param options
+ */
+export async function exportCommunityResidentExcel(
+  exportId?: number,
   options?: Record<string, any>,
 ) {
-  return request.post<API.ResponseSuccess & API.ResponseException>(communityResidentImportExcel, {
-    requestType: 'form',
-    data: formData,
+  return request.get<API.ExportFileProgress & API.ResponseException>(communityResidentExportExcel, {
+    params: {
+      exportId,
+    },
     ...(options || {}),
   });
 }
 
-/** 下载表格 */
-export async function downloadCommunityResidentExcel(options?: Record<string, any>) {
+/**
+ * 下载表格
+ * @param exportId
+ * @param options
+ */
+export async function downloadCommunityResidentExcel(
+  exportId: number,
+  options?: Record<string, any>,
+) {
   return request.get(communityResidentDownloadExcel, {
+    params: {
+      exportId,
+    },
     getResponse: true,
     responseType: 'blob',
     ...(options || {}),
